小改; 去掉不对外开放的菜单

+++ src/main/java/com/aiyi/game/dnfserver/controller/AccountController.java
@@ -69,7 +69,7 @@ public class AccountController {
      * @return
      */
     @GetMapping("{uid}")
-    public AccountVO info(@PathVariable int uid){
+    public AccountVO info(@PathVariable long uid){
         return accountService.info(uid);
     }
